Ocean and fish
Ocean when initialized will create fish in order to manage. Fish are stored in an array and when a fisherman tries to harvest a fish from an ocean node it will take the front most fish.

Fish will try to avoid being fished by swapping with fish in higher positions